Added Id validation and updated error messages

+++ routes/api/notifications.js
@@ -3,7 +3,7 @@ const express = require("express");
 const mongoose = require('mongoose')
 const router = express.Router();
 const validator = require('../../validations/notificationValidations')
-
+const idValidator = require("validator");
 // Models
 const Notification = require("../../models/Notification");
 
@@ -17,15 +17,23 @@ router.get('/', async (req,res) => {
 router.get("/:id",async (req, res) => {
   const notificationID = req.params.id;
   
-  try{
-    const notification = await Notification.findById(notificationID)
-    res.json({ data: notification})
+  if(idValidator.isMongoId(notificationID)){
+    try{
+      const notification = await Notification.findById(notificationID)
+      if(notification)
+        res.json({ data: notification})
+      else
+        res.json({msg:"Id Not Found"})
+    }
+    catch(error) {
+       res.json({ error: "ErrorOcurred"});
+    }
+
   }
-  catch(error) {
-    // We will be handling the error later
+  else{
     res.json({ error: "InvalidID"});
-   
-}  
+  }
+  
 
   
 });
@@ -42,7 +50,7 @@ router.post('/', async (req,res) => {
     res.json({msg:'Notification was created successfully', data: newNotification})
   }
   catch(error) {
-      console.log(error)
+    res.json({msg:"Error Ocurred"})
   }  
 })
 
@@ -51,39 +59,52 @@ router.post('/', async (req,res) => {
 
 // Update a notification
 router.put("/:id", async (req,res) => {
-  try {
-
-    const id = req.params.id;
+  const id = req.params.id;
+ 
+  if(!idValidator.isMongoId(id))
+    res.json({msg:"Invalid ID"});
+  else
+  {
+    try {
+      const isValidated = validator.updateValidation(req.body)
+      if (isValidated.error) return res.status(400).send({ error: isValidated.error.details[0].message })
+      
+     const oldNotification = await Notification.findByIdAndUpdate(id,req.body)
     
-   
-    const isValidated = validator.updateValidation(req.body)
-    console.log(isValidated.error)
-    if (isValidated.error) return res.status(400).send({ error: isValidated.error.details[0].message })
+     if(!oldNotification) return res.status(404).send({error: 'notification does not exist'})
+     const newNotification =await Notification.findById(id)
     
-   const oldNotification = await Notification.findByIdAndUpdate(id,req.body)
-  
-   if(!oldNotification) return res.status(404).send({error: 'notification does not exist'})
-   const newNotification =await Notification.findById(id)
-  
-   res.json({msg: 'Notification updated successfully',data: newNotification})
+     res.json({msg: 'Notification updated successfully',data: newNotification})
+    }
+    catch(error) {
+      res.json({msg:"Error Ocurred"})
+    }  
   }
-  catch(error) {
-      // We will be handling the error later
-      console.log(error)
-  }  
+
+ 
 })
 
 // delete a notification
 router.delete("/:id", async (req,res) => {
-  try {
-    const notificationID = req.params.id;
-    const deletedNotification = await Notification.findByIdAndRemove(notificationID)
-    res.json({msg:'Notification was deleted successfully', data: deletedNotification})
+  const notificationID = req.params.id;
+  if(!idValidator.isMongoId(notificationID))
+    res.json({msg:"Invalid ID"});
+  else
+  {
+    try {
+   
+      const deletedNotification = await Notification.findByIdAndRemove(notificationID)
+      if(deletedNotification)
+          res.json({msg:'Notification was deleted successfully', data: deletedNotification})
+      else
+        res.json({error:'notification does not exist'});
+    }
+    catch(error) {
+      //  console.log("error")
+       res.json({msg:"Error Ocurred"})
+    }
   }
-  catch(error) {
-     console.log(error)
-     res.json({msg:"Error Ocurred"})
-  }  
+    
 })
 
 module.exports = router;
